requirements.txt
requests==2.17.3
pycrypto==2.6.1
pymongo==3.10.0